backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
psycopg2-binary
SQLAlchemy
